last-version: Add capability to test failures
There was a bug discovered in #12 in which case the command should have
failed but it didn't. To be able to test this sort of situations, we
need to be able to test for failures.

Now tests can be declared to expect an ExitError. To make everything
clearer, also make the normal main() function not return anything, and
just report failures via exceptions. This also removed the need to
monkey-patch the die() function.