last-version: Fix error when splitting
When splitting the `apt-cache madison` output, by specifying the ' '
character to split on, it makes split to not use the "feature" to pack
multiple spaces together (or put another way, to discard empty strings in
the results).

Since some output have more than one space to separate the version, the
split was shifted and the version empty.

A test case with dmd-bin real output is used, as this problem was triggered
by a typo in dmd-bin version. See #12 for details.